Audit and harden package: fix mypy, add py.typed, 66 new tests
- Add py.typed marker for PEP 561 compliance
- Configure mypy for namespace packages (mypy_path, namespace_packages)
- Fix all mypy strict errors (optional temperature types, unused ignores)
- Add 66 comprehensive tests covering:
  - Template substitution (nested dicts, lists, edge cases)
  - JSON path extraction (dot notation, list indices, errors)
  - Header env var expansion
  - GenericHTTPAdapter (request/response, headers, auth, temperature)
  - Optional temperature in LLM adapters
  - Cache key with URL and None temperature
  - Config parsing for all new fields
  - Pre-flight cost estimation (known/unknown models, sequential stopping)
  - Cost/reliability arbitrage table
  - Post-run cost with cost_per_request
  - CLI --cost-per-request and --yes flags
- 398 tests passing, ruff clean, mypy strict clean

+++ src/theaios/trustgate/certification.py
@@ -201,7 +201,7 @@ def estimate_cost(
 
     # Use user-provided cost_per_request if available
     if config.endpoint.cost_per_request is not None:
-        return n_calls * config.endpoint.cost_per_request
+        return float(n_calls * config.endpoint.cost_per_request)
 
     # Fall back to known model pricing
     model = config.endpoint.model
